+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.GroceryApp"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.App.MainActivity"
+            android:name=".presentation.app.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.GroceryApp">
