+++ app/src/main/java/com/example/groceryapp/presentation/cart/CartScreen.kt
@@ -1,4 +1,4 @@
-package com.example.groceryapp.presentation.Explore
+package com.example.groceryapp.presentation.cart
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
@@ -7,10 +7,10 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
-import com.example.groceryapp.presentation.BottomNavigationBar.BottomNavigation
+import com.example.groceryapp.presentation.bottomNavigationBar.BottomNavigation
 
 @Composable
-fun ExploreScreen(navController: NavController) {
+fun CartScreen(navController: NavController) {
 
     Scaffold(
         bottomBar = { BottomNavigation(navController) },
