+++ app/src/main/java/com/example/groceryapp/presentation/favorite/FavoriteScreen.kt
@@ -1,4 +1,4 @@
-package com.example.groceryapp.presentation.Favorite
+package com.example.groceryapp.presentation.favorite
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
@@ -7,7 +7,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
-import com.example.groceryapp.presentation.BottomNavigationBar.BottomNavigation
+import com.example.groceryapp.presentation.bottomNavigationBar.BottomNavigation
 
 @Composable
 fun FavoriteScreen(navController: NavController) {
